getMemorySize works on 32-bit by returning 64-bit integer always

+++ timeutil.h
@@ -194,7 +194,7 @@ inline uint64_t getTotalSystemMemory()
 /**
  * Returns the size of physical memory (RAM) in bytes.
  */
-inline size_t getMemorySize( )
+inline uint64_t getMemorySize( )
 {
 #if defined(_WIN32) && (defined(__CYGWIN__) || defined(__CYGWIN32__))
 	/* Cygwin under Windows. ------------------------------------ */
@@ -202,7 +202,7 @@ inline size_t getMemorySize( )
 	MEMORYSTATUS status;
 	status.dwLength = sizeof(status);
 	GlobalMemoryStatus( &status );
-	return (size_t)status.dwTotalPhys;
+	return (uint64_t)status.dwTotalPhys;
 
 #elif defined(_WIN32)
 	/* Windows. ------------------------------------------------- */
@@ -210,7 +210,7 @@ inline size_t getMemorySize( )
 	MEMORYSTATUSEX status;
 	status.dwLength = sizeof(status);
 	GlobalMemoryStatusEx( &status );
-	return (size_t)status.ullTotalPhys;
+	return (uint64_t)status.ullTotalPhys;
 
 #elif defined(__unix__) || defined(__unix) || defined(unix) || (defined(__APPLE__) && defined(__MACH__))
 	/* UNIX variants. ------------------------------------------- */
@@ -224,25 +224,25 @@ inline size_t getMemorySize( )
 #elif defined(HW_PHYSMEM64)
 	mib[1] = HW_PHYSMEM64;          /* NetBSD, OpenBSD. --------- */
 #endif
-	int64_t size = 0;               /* 64-bit */
+	uint64_t size = 0;               /* 64-bit */
 	size_t len = sizeof( size );
 	if ( sysctl( mib, 2, &size, &len, NULL, 0 ) == 0 )
-		return (size_t)size;
+		return (uint64_t)size;
 	return 0L;			/* Failed? */
 
 #elif defined(_SC_AIX_REALMEM)
 	/* AIX. ----------------------------------------------------- */
-	return (size_t)sysconf( _SC_AIX_REALMEM ) * (size_t)1024L;
+	return (uint64_t)sysconf( _SC_AIX_REALMEM ) * (uint64_t)1024L;
 
 #elif defined(_SC_PHYS_PAGES) && defined(_SC_PAGESIZE)
 	/* FreeBSD, Linux, OpenBSD, and Solaris. -------------------- */
-	return (size_t)sysconf( _SC_PHYS_PAGES ) *
-		(size_t)sysconf( _SC_PAGESIZE );
+	return (uint64_t)sysconf( _SC_PHYS_PAGES ) *
+		(uint64_t)sysconf( _SC_PAGESIZE );
 
 #elif defined(_SC_PHYS_PAGES) && defined(_SC_PAGE_SIZE)
 	/* Legacy. -------------------------------------------------- */
-	return (size_t)sysconf( _SC_PHYS_PAGES ) *
-		(size_t)sysconf( _SC_PAGE_SIZE );
+	return (uint64_t)sysconf( _SC_PHYS_PAGES ) *
+		(uint64_t)sysconf( _SC_PAGE_SIZE );
 
 #elif defined(CTL_HW) && (defined(HW_PHYSMEM) || defined(HW_REALMEM))
 	/* DragonFly BSD, FreeBSD, NetBSD, OpenBSD, and OSX. -------- */
@@ -253,10 +253,10 @@ inline size_t getMemorySize( )
 #elif defined(HW_PYSMEM)
 	mib[1] = HW_PHYSMEM;		/* Others. ------------------ */
 #endif
-	unsigned int size = 0;		/* 32-bit */
+	uint64_t size = 0;		/* 32-bit */
 	size_t len = sizeof( size );
 	if ( sysctl( mib, 2, &size, &len, NULL, 0 ) == 0 )
-		return (size_t)size;
+		return (uint64_t)size;
 	return 0L;			/* Failed? */
 #endif /* sysctl and sysconf variants */
